requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
alembic==1.4.2            # via jupyterhub
appnope==0.1.0            # via ipykernel, ipython
async-generator==1.10     # via jupyterhub
attrs==19.3.0             # via jsonschema
awscli==1.18.74           # via -r requirements.in
backcall==0.1.0           # via ipython
bleach==3.1.5             # via nbconvert
botocore==1.16.24         # via awscli, s3transfer
certifi==2020.4.5.1       # via requests
certipy==0.1.3            # via jupyterhub
cffi==1.14.0              # via cryptography
chardet==3.0.4            # via requests
click==7.1.2              # via -r requirements.in
colorama==0.4.3           # via awscli
cryptography==2.9.2       # via pyopenssl
cycler==0.10.0            # via matplotlib
decorator==4.4.2          # via ipython, networkx, traitlets
defusedxml==0.6.0         # via nbconvert
docutils==0.15.2          # via awscli, botocore
entrypoints==0.3          # via jupyterhub, nbconvert
gevent==20.6.2            # via grequests
greenlet==0.4.16          # via gevent
grequests==0.6.0          # via -r requirements.in
idna==2.9                 # via requests
imageio==2.8.0            # via scikit-image
ipykernel==5.3.0          # via ipympl, ipywidgets, notebook
ipympl==0.5.6             # via -r requirements.in
ipython-genutils==0.2.0   # via nbformat, notebook, traitlets
ipython==7.15.0           # via -r requirements.in, ipykernel, ipywidgets
ipywidgets==7.5.1         # via -r requirements.in, ipympl
jedi==0.17.0              # via ipython
jinja2==2.11.2            # via jupyterhub, jupyterlab, jupyterlab-server, nbconvert, notebook
jmespath==0.10.0          # via botocore
joblib==0.15.1            # via scikit-learn
json5==0.9.5              # via jupyterlab-server
jsonschema==3.2.0         # via jupyter-telemetry, jupyterlab-server, nbformat
jupyter-client==6.1.3     # via ipykernel, notebook
jupyter-core==4.6.3       # via jupyter-client, nbconvert, nbformat, notebook
jupyter-telemetry==0.1.0  # via jupyterhub
jupyterhub==1.1.0         # via -r requirements.in
jupyterlab-server==1.1.5  # via jupyterlab
jupyterlab==2.1.4         # via -r requirements.in
kiwisolver==1.2.0         # via matplotlib
mako==1.1.3               # via alembic
markupsafe==1.1.1         # via jinja2, mako
matplotlib==3.2.1         # via -r requirements.in, ipympl, scikit-image, seaborn
mistune==0.8.4            # via nbconvert
nbconvert==5.6.1          # via notebook
nbformat==5.0.6           # via ipywidgets, nbconvert, notebook
networkx==2.4             # via scikit-image
notebook==6.0.3           # via -r requirements.in, jupyterlab, jupyterlab-server, widgetsnbextension
numpy==1.18.5             # via imageio, matplotlib, pandas, patsy, pywavelets, scikit-learn, scipy, seaborn, statsmodels
oauthlib==3.1.0           # via jupyterhub
packaging==20.4           # via bleach
pamela==1.0.0             # via jupyterhub
pandas==1.0.4             # via -r requirements.in, seaborn, statsmodels
pandocfilters==1.4.2      # via nbconvert
parso==0.7.0              # via jedi
patsy==0.5.1              # via statsmodels
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
pillow==7.1.2             # via imageio, scikit-image
plotly==4.8.1             # via -r requirements.in
prometheus-client==0.8.0  # via jupyterhub, notebook
prompt-toolkit==3.0.5     # via ipython
ptyprocess==0.6.0         # via pexpect, terminado
pyasn1==0.4.8             # via rsa
pycparser==2.20           # via cffi
pygments==2.6.1           # via ipython, nbconvert
pyopenssl==19.1.0         # via certipy
pyparsing==2.4.7          # via matplotlib, packaging
pyrsistent==0.16.0        # via jsonschema
python-dateutil==2.8.1    # via alembic, botocore, jupyter-client, jupyterhub, matplotlib, pandas
python-dotenv==0.13.0     # via -r requirements.in
python-editor==1.0.4      # via alembic
python-json-logger==0.1.11  # via jupyter-telemetry
pytz==2020.1              # via pandas
pywavelets==1.1.1         # via scikit-image
pyyaml==5.3.1             # via awscli
pyzmq==19.0.1             # via jupyter-client, notebook
requests==2.23.0          # via -r requirements.in, grequests, jupyterhub, jupyterlab-server
retrying==1.3.3           # via plotly
rsa==3.4.2                # via awscli
ruamel.yaml.clib==0.2.0   # via ruamel.yaml
ruamel.yaml==0.16.10      # via jupyter-telemetry
s3transfer==0.3.3         # via awscli
scikit-image==0.16.2      # via -r requirements.in
scikit-learn==0.22.2.post1  # via -r requirements.in
scipy==1.4.1              # via scikit-image, scikit-learn, seaborn, statsmodels
seaborn==0.10.1           # via -r requirements.in
send2trash==1.5.0         # via notebook
six==1.15.0               # via bleach, cryptography, cycler, jsonschema, packaging, patsy, plotly, pyopenssl, pyrsistent, python-dateutil, retrying, traitlets
sqlalchemy==1.3.17        # via alembic, jupyterhub
statsmodels==0.11.1       # via -r requirements.in
terminado==0.8.3          # via notebook
testpath==0.4.4           # via nbconvert
tornado==6.0.4            # via ipykernel, jupyter-client, jupyterhub, jupyterlab, notebook, terminado
traitlets==4.3.3          # via ipykernel, ipython, ipywidgets, jupyter-client, jupyter-core, jupyter-telemetry, jupyterhub, nbconvert, nbformat, notebook
urllib3==1.25.9           # via botocore, requests
wcwidth==0.2.3            # via prompt-toolkit
webencodings==0.5.1       # via bleach
widgetsnbextension==3.5.1  # via ipywidgets
zope.event==4.4           # via gevent
zope.interface==5.1.0     # via gevent

# The following packages are considered to be unsafe in a requirements file:
# setuptools
